.github/requirements.txt
moz-fluent-linter==0.4.*
moz-l10n==0.6.0
